IOS changes to test out storyboard
I am playing around with storyboard in order to make sure that it is
able to fit the needs of the application. No actual functionality has
been put it in.

I have added in my own custom ViewController and a number of screens.
In the View Controller i implemented some functionality that
corresponds wit she omf othe ui elements.